+++ rlzoo/algorithms/sac/sac.py
@@ -62,7 +62,6 @@ class SAC():
 
     def evaluate(self, state, epsilon=1e-6):
         ''' generate action with state for calculating gradients '''
-        state = state.astype(np.float32)
         _ = self.policy_net(state)
         mean, log_std = self.policy_net.policy_dist.get_param()  # as SAC uses TanhNorm instead of normal distribution, need original mean_std
         std = tf.math.exp(log_std)  # no clip in evaluation, clip affects gradients flow
@@ -83,11 +82,11 @@ class SAC():
 
     def get_action(self, state):
         ''' generate action with state for interaction with envronment '''
-        return self.policy_net(np.array([state.astype(np.float32)])).numpy()[0]
+        return self.policy_net(np.array([state])).numpy()[0]
 
     def get_action_greedy(self, state):
         ''' generate action with state for interaction with envronment '''
-        return self.policy_net(np.array([state.astype(np.float32)]), greedy=True).numpy()[0]
+        return self.policy_net(np.array([state]), greedy=True).numpy()[0]
 
     def sample_action(self, ):
         ''' generate random actions for exploration '''
@@ -211,7 +210,6 @@ class SAC():
             t0 = time.time()
             for eps in range(train_episodes):
                 state = env.reset()
-                state = state.astype(np.float32)
                 episode_reward = 0
 
                 for step in range(max_steps):
@@ -221,7 +219,6 @@ class SAC():
                         action = self.sample_action()
 
                     next_state, reward, done, _ = env.step(action)
-                    next_state = next_state.astype(np.float32)
                     if render: env.render()
                     done = 1 if done == True else 0
 
@@ -264,13 +261,11 @@ class SAC():
 
             for eps in range(test_episodes):
                 state = env.reset()
-                state = state.astype(np.float32)
                 episode_reward = 0
 
                 for step in range(max_steps):
                     action = self.get_action_greedy(state)
                     next_state, reward, done, _ = env.step(action)
-                    next_state = next_state.astype(np.float32)
                     if render: env.render()
                     done = 1 if done == True else 0
